+++ .gitignore
@@ -34,6 +34,7 @@ local.properties
 node_modules/
 npm-debug.log
 yarn-error.log
+yarn.lock
 
 # BUCK
 buck-out/
@@ -58,5 +59,4 @@ buck-out/
 # CocoaPods
 /ios/Pods/
 
-#lock
-yarn.lock
+
