+++ index.js
@@ -7,7 +7,6 @@ import App from './src/navigator/App';
 import Search from './src/components/Search';
 import Home from './src/components/Home';
 import AppNavi from './src/navigator/AppNavigation';
-import HomeS from './src/components/HomeComponent';
 import {name as appName} from './app.json';
 
 AppRegistry.registerComponent(appName, () => AppNavi);
